# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IRV Intelligent Road Vision",
    "description": "车牌识别 + 手势控车 + Agent 预警 + 沙盘摄像头统一后端",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login/password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login Password",
        "operationId": "login_password_api_auth_login_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/sms/send": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Sms Send",
        "operationId": "sms_send_api_auth_sms_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmsSendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login/sms": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login Sms",
        "operationId": "login_sms_api_auth_login_sms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmsLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/email/send": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Email Send",
        "operationId": "email_send_api_auth_email_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailSendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login/email": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login Email",
        "operationId": "login_email_api_auth_login_email_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmailLoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "code": {
            "type": "string",
            "maxLength": 8,
            "minLength": 4,
            "title": "Code"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": "driver"
          }
        },
        "type": "object",
        "required": [
          "email",
          "code"
        ],
        "title": "EmailLoginRequest"
      },
      "EmailSendRequest": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 128,
            "minLength": 5,
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": "driver"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "EmailSendRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PasswordLoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "PasswordLoginRequest"
      },
      "RegisterRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 64,
            "minLength": 2,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 6,
            "title": "Password"
          },
          "displayName": {
            "type": "string",
            "maxLength": 128,
            "title": "Displayname",
            "default": ""
          },
          "contact": {
            "type": "string",
            "maxLength": 128,
            "title": "Contact",
            "default": ""
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": "driver"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "RegisterRequest"
      },
...
      "SmsLoginRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "code": {
            "type": "string",
            "maxLength": 8,
            "minLength": 4,
            "title": "Code"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": "driver"
          }
        },
        "type": "object",
        "required": [
          "phone",
          "code"
        ],
        "title": "SmsLoginRequest"
      },
      "SmsSendRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "maxLength": 32,
            "minLength": 5,
            "title": "Phone"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": "driver"
          }
        },
        "type": "object",
        "required": [
          "phone"
        ],
        "title": "SmsSendRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}